src/detachdragzone.cpp: fix undefined behaviour
Use qobject_cast instead of static cast to detect if source is of type TooMenuItem or not. If it is a ToolMenuItem only then we can detach.
As ToolMenuItem inherits from QObject, and all MenuItems will inherit from QObject, we can use qobject_cast instead of dynamic_cast, which does not use RTTI to check if the cast is valid.

Signed-off-by: Daniel Guramulta <daniel.guramulta@analog.com>