Voltmeter: fix GUI after editing .ui file
There seem to be some properties manually added to the .ui xml file that qt creator removes after each time we touch the .ui file using the Qt designer. This is annoying and for sure not the right way of doing things. Let's programmatically set these properties so we don't need to manually edit the .ui file after each time we open it in qt designer

Signed-off-by: DanielGuramulta <Daniel.Guramulta@analog.com>